Refactor NetworkWriteService to support Non-MD meters with UINT32 Port fix
- Updated writeApn to use OBIS 0.11.25.4.0.255 for Non-MD meters.
- Updated writeIpPort to support Non-MD meters by parsing "IP:Port" and writing Port (Class 41, UINT32) then IP (Class 45, Octet String) in sequence.
- UINT32 (Tag 0x06) used for Port to ensure hardware compatibility as per diagnostic feedback.
- Injected MeterRepository into NetworkWriteService for meter class lookup.
- Enhanced NetworkWriteServiceTest with MD and Non-MD test cases verifying OBIS and Data Types.

+++ hes/src/main/java/com/memmcol/hes/domain/network/NetworkWriteService.java
@@ -112,8 +112,9 @@ public class NetworkWriteService {
                 log.info("Writing Port {} then IP {} to Non-MD meter {}", port, ip, meterSerial);
 
                 // Fallback: Write Port first (Class 41, Attr 2), then IP (Class 45, Attr 5)
+                // Using UINT32 (double-long-unsigned) for Port to ensure hardware compatibility
                 response = dlmsReaderUtils.writeAttribute(client, meterSerial, "0.11.25.0.0.255", 41, 2,
-                        port, DataType.UINT16);
+                        (long) port, DataType.UINT32);
 
                 if (response.isSuccess()) {
                     response = dlmsReaderUtils.writeAttribute(client, meterSerial, "0.11.25.4.0.255", 45, 5,
